requirements.txt
aiohttp
matplotlib
flask
numpy
flask-swagger-ui
PyYAML
asyncio
